+++ CHANGELOG.md
@@ -4,13 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.2.4] - 2019-04-02
+## Unreleased
 
 ### auth0-github-deploy v2.7.2
 ### auth0-visualstudio-deploy v2.6.3
+### auth0-bitbucket-deploy v2.7.3
 
-- #### Fixed
-  - Occasional `request-promise` issues fixed.
+- #### Changed
+  - `request-promise` is replaced by `axios`.
   
 ## [1.2.3] - 2019-03-25